@@ -1,6 +1,6 @@
 {
   "name": "keycloak-angular-ui",
-  "version": "5.0.0",
+  "version": "6.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --watch",
@@ -11,31 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.1",
-    "@angular/common": "^13.3.1",
-    "@angular/compiler": "^13.3.1",
-    "@angular/core": "^13.3.1",
-    "@angular/forms": "^13.3.1",
-    "@angular/platform-browser": "^13.3.1",
-    "@angular/platform-browser-dynamic": "^13.3.1",
-    "@angular/router": "^13.3.1",
+    "@angular/animations": "^14.0.0",
+    "@angular/common": "^14.0.0",
+    "@angular/compiler": "^14.0.0",
+    "@angular/core": "^14.0.0",
+    "@angular/forms": "^14.0.0",
+    "@angular/platform-browser": "^14.0.0",
+    "@angular/platform-browser-dynamic": "^14.0.0",
+    "@angular/router": "^14.0.0",
     "bootstrap": "^5.1.3",
     "jquery": "^3.6.0",
-    "keycloak-js": "^16.1.1",
-    "ngx-spinner": "^12.0.0",
+    "keycloak-js": "^18.0.0",
+    "ngx-spinner": "^13.1.1",
     "popper.js": "^1.16.1",
-    "rxjs": "~7.4.0",
+    "rxjs": "~7.5.5",
     "tslib": "^2.3.1",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.1",
-    "@angular/cli": "^13.3.1",
-    "@angular/compiler-cli": "^13.3.1",
-    "@angular/language-service": "^13.3.1",
+    "@angular-devkit/build-angular": "14.0.0",
+    "@angular/cli": "^14.0.0",
+    "@angular/compiler-cli": "^14.0.0",
+    "@angular/language-service": "^14.0.0",
     "@types/jasmine": "~3.5.14",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "^12.20.36",
+    "@types/node": "^17.0.39",
     "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -47,6 +47,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.0",
-    "typescript": "^4.6.3"
+    "typescript": "^4.7.3"
   }
 }
